cemetery background [unfinished version]
changed world border, dunno if that's a problem
also i drag and dropped sprites into the assets folder now there are question marks on them, also dunno if that'S a problem. if i did something wrong, please tell me so i can avoid that in the future
also, is it normal that i can't see the player character?